refactor: move map_size function to twomap module
